Restructuring and Backbone implementation changes
