+++ keras_cv/src/models/object_detection/mask_rcnn/roi_sampler.py
